Moved redcap views into a blueprint
