+++ dashboard/redcap/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+rcap_bp = Blueprint('redcap', __name__)
+
+from . import views
